Decided on the method for inserting IDs (including Member ID and Circle ID) into the database (#19)
* i32 -> u16

* owner

* delete as

+++ src/handler.rs
@@ -67,10 +67,10 @@ pub async fn handle_get_test(State(state): State<AppState>) -> impl IntoResponse
 #[derive(Debug, serde::Deserialize, serde::Serialize)]
 pub struct CreateCircleRequestBody {
     pub circle_name: String,
-    pub capacity: i32,
+    pub capacity: u16,
     pub owner_name: String,
-    pub owner_age: i32,
-    pub owner_grade: i32,
+    pub owner_age: u16,
+    pub owner_grade: u16,
     pub owner_major: String,
 }
 
@@ -98,8 +98,8 @@ impl std::convert::From<CreateCircleRequestBody> for CreateCircleInput {
 
 #[derive(Debug, serde::Deserialize, serde::Serialize)]
 pub struct CreateCircleResponseBody {
-    pub circle_id: i32,
-    pub owner_id: i32,
+    pub circle_id: u16,
+    pub owner_id: u16,
 }
 
 impl std::convert::From<CreateCircleOutput> for CreateCircleResponseBody {
@@ -132,14 +132,14 @@ pub async fn handle_create_circle(
 
 #[derive(Debug, Deserialize)]
 pub struct FetchCircleInputParam {
-    id: i32,
+    id: u16,
 }
 
 #[derive(Debug, serde::Deserialize, serde::Serialize)]
 pub struct FetcheCircleResponseBody {
-    pub circle_id: i32,
+    pub circle_id: u16,
     pub circle_name: String,
-    pub capacity: i32,
+    pub capacity: u16,
     pub owner: MemberOutput,
     pub members: Vec<MemberOutput>,
 }
@@ -180,24 +180,24 @@ pub async fn handle_fetch_circle(
 
 #[derive(Debug, Deserialize)]
 pub struct UpdateCircleInputParam {
-    id: i32,
+    id: u16,
 }
 
 #[derive(Debug, serde::Deserialize, serde::Serialize)]
 pub struct UpdateCircleRequestBody {
     pub circle_name: Option<String>,
-    pub capacity: Option<i32>,
+    pub capacity: Option<u16>,
 }
 
 impl UpdateCircleRequestBody {
-    pub fn convert_to_input(self, id: i32) -> UpdateCircleInput {
+    pub fn convert_to_input(self, id: u16) -> UpdateCircleInput {
         UpdateCircleInput::new(id, self.circle_name, self.capacity)
     }
 }
 
 #[derive(Debug, serde::Serialize)]
 pub struct UpdateCircleResponseBody {
-    pub id: i32,
+    pub id: u16,
 }
 
 impl std::convert::From<UpdateCircleOutPut> for UpdateCircleResponseBody {
