+++ src/usecase/create_circle.rs
@@ -13,20 +13,20 @@ use crate::domain::{
 #[derive(Debug, Deserialize)]
 pub struct CreateCircleInput {
     pub circle_name: String,
-    pub capacity: i32,
+    pub capacity: u16,
     pub owner_name: String,
-    pub owner_age: i32,
-    pub owner_grade: i32,
+    pub owner_age: u16,
+    pub owner_grade: u16,
     pub owner_major: String,
 }
 
 impl CreateCircleInput {
     pub fn new(
         circle_name: String,
-        capacity: i32,
+        capacity: u16,
         owner_name: String,
-        owner_age: i32,
-        owner_grade: i32,
+        owner_age: u16,
+        owner_grade: u16,
         owner_major: String,
     ) -> Self {
         CreateCircleInput {
@@ -42,8 +42,8 @@ impl CreateCircleInput {
 
 #[derive(Debug, Deserialize)]
 pub struct CreateCircleOutput {
-    pub circle_id: i32,
-    pub owner_id: i32,
+    pub circle_id: u16,
+    pub owner_id: u16,
 }
 
 pub struct CreateCircleUsecase<T>
@@ -85,8 +85,8 @@ where
             .create(&circle)
             .await
             .map(|_| CreateCircleOutput {
-                circle_id: i32::from(circle.id),
-                owner_id: i32::from(owner_id),
+                circle_id: u16::from(circle.id),
+                owner_id: u16::from(owner_id),
             })
     }
 }
